feat: add requireAuthenticated middleware
- feat: add checkValidation middleware
